requirements.txt
opencv-python
numpy
torch
torchvision
Pillow
matplotlib
fpdf
facenet-pytorch
tk
